+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Improve commit behavior
+- Move gg commit conventions from gg_git to gg_one_core
 
 ## 2.2.0 - 2026-08-09